REFACTOR `changelog()` & `versionChanges()` methods to ChangelogTest
